remove some VinylDNSTestData usage
