+++ modules/api/src/test/scala/vinyldns/api/VinylDNSTestData.scala
@@ -16,11 +16,8 @@
 
 package vinyldns.api
 
-import java.util.UUID
-
 import org.joda.time.DateTime
-import vinyldns.api.domain.record.{ListRecordSetChangesResponse, RecordSetChangeGenerator}
-import vinyldns.api.domain.zone._
+import vinyldns.api.domain.record.RecordSetChangeGenerator
 import vinyldns.api.repository.TestDataLoader
 import vinyldns.core.TestRecordSetData
 import vinyldns.core.domain.auth.AuthPrincipal
@@ -44,24 +41,18 @@ trait VinylDNSTestData {
     ZoneConnection("vinyldns.", "vinyldns.", "nzisn+4G2ldMn0q1CV3vsg==", "10.1.1.1"))
 
   val okZone: Zone = Zone("ok.zone.recordsets.", "test@test.com", adminGroupId = grp.id)
-  val zoneDeleted: Zone = okZone.copy(status = ZoneStatus.Deleted)
   val zoneActive: Zone = Zone(
     "ok.zone.recordsets.",
     "test@test.com",
     status = ZoneStatus.Active,
     connection = testConnection)
 
-  val zoneAuthorized: Zone = zoneActive.copy(adminGroupId = grp.id)
-  val zoneNotAuthorized: Zone =
-    zoneActive.copy(id = UUID.randomUUID().toString, adminGroupId = UUID.randomUUID().toString)
   val zoneIp4: Zone = okZone.copy(name = "0.162.198.in-addr.arpa.")
   val zoneIp6: Zone = okZone.copy(name = "1.9.e.f.c.c.7.2.9.6.d.f.ip6.arpa.")
 
   val rsOk: RecordSet = TestRecordSetData.rsOk
 
   val aaaa: RecordSet = TestRecordSetData.aaaa
-
-  val validAAAAToOrigin: RecordSet = TestRecordSetData.aaaaOrigin
 
   val cname: RecordSet = TestRecordSetData.cname
 
@@ -89,31 +80,9 @@ trait VinylDNSTestData {
 
   val ptrIp6: RecordSet = TestRecordSetData.ptrIp6
 
-  val soa: RecordSet = RecordSet(
-    ptrIp4.name,
-    "soa.foo.bar.com",
-    RecordType.SOA,
-    200,
-    RecordSetStatus.Active,
-    DateTime.now,
-    None,
-    List(SOAData("something", "other", 1, 2, 3, 5, 6)))
-
   val srv: RecordSet = TestRecordSetData.srv
 
   val ns: RecordSet = TestRecordSetData.ns
-
-  val mx: RecordSet = TestRecordSetData.mx
-
-  val txt: RecordSet = RecordSet(
-    okZone.id,
-    "txt",
-    RecordType.TXT,
-    200,
-    RecordSetStatus.Pending,
-    DateTime.now,
-    None,
-    List(TXTData("txt")))
 
   val zoneCreate: ZoneChange = ZoneChange(
     okZone,
@@ -128,42 +97,7 @@ trait VinylDNSTestData {
     ZoneChangeStatus.Complete,
     created = DateTime.now)
 
-  val notAuthorizedZone: Zone = Zone("notAuth.", "notAuth@notAuth.com")
-  val notAuthorizedAuth: AuthPrincipal =
-    okAuth.copy(signedInUser = okAuth.signedInUser.copy(userName = "not auth", id = "not auth"))
-
   val pendingCreateAAAA: RecordSetChange = RecordSetChangeGenerator.forAdd(aaaa, zoneActive, okAuth)
-  val pendingCreateCNAME: RecordSetChange =
-    RecordSetChangeGenerator.forAdd(cname, zoneActive, okAuth)
-  val pendingChangeSet: ChangeSet = ChangeSet(Seq(pendingCreateAAAA, pendingCreateCNAME))
-  val pendingCreateNS: RecordSetChange = RecordSetChangeGenerator.forAdd(ns, zoneActive, okAuth)
-
-  val completeCreateAAAA: RecordSetChange =
-    pendingCreateAAAA.copy(status = RecordSetChangeStatus.Complete)
-  val completeCreateCNAME: RecordSetChange =
-    pendingCreateCNAME.copy(status = RecordSetChangeStatus.Complete)
-  val completeChangeSet: ChangeSet = ChangeSet(Seq(completeCreateAAAA, completeCreateCNAME))
-  val completeCreateNS: RecordSetChange =
-    pendingCreateNS.copy(status = RecordSetChangeStatus.Complete)
-
-  val listZoneChangesResponse: ListZoneChangesResponse = ListZoneChangesResponse(
-    zoneActive.id,
-    List(zoneCreate.copy(id = zoneActive.id)),
-    nextId = None,
-    startFrom = None,
-    maxItems = 100)
-  val changesWithUserName: List[RecordSetChangeInfo] = List(
-    RecordSetChangeInfo(pendingCreateAAAA, Some("ok")),
-    RecordSetChangeInfo(pendingCreateCNAME, Some("ok")),
-    RecordSetChangeInfo(completeCreateAAAA, Some("ok")),
-    RecordSetChangeInfo(completeCreateCNAME, Some("ok"))
-  )
-  val listRecordSetChangesResponse: ListRecordSetChangesResponse = ListRecordSetChangesResponse(
-    zoneActive.id,
-    changesWithUserName,
-    nextId = None,
-    startFrom = None,
-    maxItems = 100)
 
   val zoneChangePending: ZoneChange =
     ZoneChange(okZone, "ok", ZoneChangeType.Update, ZoneChangeStatus.Pending)
@@ -179,36 +113,6 @@ trait VinylDNSTestData {
     List(AData("10.1.1.1")))
 
   val fakeTime: DateTime = new DateTime(2010, 1, 1, 0, 0)
-
-  val userAclRule: ACLRule = ACLRule(
-    AccessLevel.Read,
-    Some("desc"),
-    Some("johnny"),
-    None,
-    Some("www-*"),
-    Set(RecordType.A, RecordType.AAAA, RecordType.CNAME))
-  val userAclRuleInfo: ACLRuleInfo = ACLRuleInfo(
-    AccessLevel.Read,
-    Some("desc"),
-    Some("johnny"),
-    None,
-    Some("www-*"),
-    Set(RecordType.A, RecordType.AAAA, RecordType.CNAME))
-  val groupAclRule: ACLRule = ACLRule(
-    AccessLevel.Read,
-    Some("desc"),
-    None,
-    Some("group"),
-    Some("www-*"),
-    Set(RecordType.A, RecordType.AAAA, RecordType.CNAME))
-  val groupAclRuleInfo: ACLRule = ACLRule(
-    AccessLevel.Read,
-    Some("desc"),
-    None,
-    Some("group"),
-    Some("www-*"),
-    Set(RecordType.A, RecordType.AAAA, RecordType.CNAME))
-  val zoneAcl: ZoneACL = ZoneACL(Set(userAclRule, groupAclRule))
 
   val baseAclRuleInfo: ACLRuleInfo =
     ACLRuleInfo(AccessLevel.Read, Some("desc"), None, Some("group"), None, Set.empty)
@@ -226,9 +130,6 @@ trait VinylDNSTestData {
       zone = anonymizeTimeOnly(recordSetChange.zone)
     )
 
-  def anonymize(zoneChange: ZoneChange): ZoneChange =
-    zoneChange.copy(id = "a", created = fakeTime, zone = anonymizeTimeOnly(zoneChange.zone))
-
   def anonymize(changeSet: ChangeSet): ChangeSet =
     changeSet.copy(
       id = "a",
@@ -238,19 +139,6 @@ trait VinylDNSTestData {
         .map(anonymize)
         .sortBy(rs => (rs.recordSet.name, rs.recordSet.typ))
     )
-
-  def anonymize(zoneDiff: ZoneDiff): ZoneDiff =
-    zoneDiff.copy(
-      anonymize(zoneDiff.zoneChange),
-      zoneDiff.recordSetChanges
-        .map(anonymize)
-        .sortBy(rs => (rs.recordSet.name, rs.recordSet.typ)))
-
-  def anonymize(zone: Zone): Zone = {
-    val newUpdate = zone.updated.map(_ => fakeTime)
-    val newLatestSync = zone.latestSync.map(_ => fakeTime)
-    zone.copy(id = "a", created = fakeTime, updated = newUpdate, latestSync = newLatestSync)
-  }
 
   def anonymizeTimeOnly(zone: Zone): Zone = {
     val newUpdate = zone.updated.map(_ => fakeTime)
